added cast
git-svn-id: svn://136.177.114.72/svn_GW/IPhreeqcMMS/trunk@7489 1feff8c3-07ed-0310-ac33-dd36852eb9cd

+++ src/IPhreeqcMMSLib.cpp
@@ -38,7 +38,7 @@ IPhreeqcMMSLib::CreateIPhreeqcMMS(void)
 	try
 	{
 		IPhreeqcMMS* IPhreeqcPtr = new IPhreeqcMMS;
-		n = IPhreeqcPtr->Index;
+		n = static_cast<int>(IPhreeqcPtr->Index);
 	}
 	catch(...)
 	{
